split e2e test files
